requirements.txt
fastapi
uvicorn
numpy
liboqs-python